@@ -6,11 +6,11 @@
 
   <groupId>com.transformuk.hee</groupId>
   <artifactId>profile-api</artifactId>
-  <version>2.4.5</version>
+  <version>2.4.6</version>
   <packaging>jar</packaging>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>11</java.version>
     <maven.compiler.source>${java.version}</maven.compiler.source>
     <maven.compiler.target>${java.version}</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
