@@ -28,14 +28,12 @@
     <commons-lang.version>3.12.0</commons-lang.version>
     <docker-maven-plugin.version>1.2.2</docker-maven-plugin.version>
     <frontend.maven.plugin.version>1.3</frontend.maven.plugin.version>
-    <hibernate.version>5.2.8.Final</hibernate.version>
-    <hikaricp.version>2.6.0</hikaricp.version>
-    <java.version>1.8</java.version>
+    <hibernate.version>5.3.34.Final</hibernate.version>
+    <hibernate.validator.version>5.3.6.Final</hibernate.validator.version>
+    <java.version>11</java.version>
     <javassist.version>3.21.0-GA</javassist.version>
     <javax.servlet.version>4.0.1</javax.servlet.version>
-    <jhipster.server.version>1.1.0</jhipster.server.version>
     <jjwt.version>0.9.1</jjwt.version>
-    <logstash-logback-encoder.version>6.6</logstash-logback-encoder.version>
     <m2e.apt.activation>jdt_apt</m2e.apt.activation>
     <mapstruct.version>1.4.2.Final</mapstruct.version>
     <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
