@@ -3,7 +3,6 @@ package com.transformuk.hee.tis.profile;
 import com.transformuk.hee.tis.audit.repository.TisAuditRepository;
 import com.transformuk.hee.tis.profile.config.ApplicationProperties;
 import com.transformuk.hee.tis.profile.config.DefaultProfileUtil;
-import io.github.jhipster.config.JHipsterConstants;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.Arrays;
@@ -13,8 +12,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.actuate.audit.AuditEventRepository;
-import org.springframework.boot.actuate.autoconfigure.MetricFilterAutoConfiguration;
-import org.springframework.boot.actuate.autoconfigure.MetricRepositoryAutoConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -30,8 +27,7 @@ import org.springframework.web.servlet.DispatcherServlet;
     "com.transformuk.hee.tis.profile",
     "com.transformuk.hee.tis.reference.client"
 })
-@EnableAutoConfiguration(exclude = {MetricFilterAutoConfiguration.class,
-    MetricRepositoryAutoConfiguration.class})
+@EnableAutoConfiguration
 @EnableConfigurationProperties({ApplicationProperties.class})
 @PropertySource(
     {
@@ -85,21 +81,18 @@ public class ProfileApp {
   /**
    * Initializes profile.
    * <p>
-   * Spring profiles can be configured with a program arguments --spring.profiles.active=your-active-profile
+   * Spring profiles can be configured with a program arguments
+   * --spring.profiles.active=your-active-profile
    * <p>
-   * You can find more information on how profiles work with JHipster on <a
-   * href="http://jhipster.github.io/profiles/">http://jhipster.github.io/profiles/</a>.
    */
   @PostConstruct
   public void initApplication() {
     Collection<String> activeProfiles = Arrays.asList(env.getActiveProfiles());
-    if (activeProfiles.contains(JHipsterConstants.SPRING_PROFILE_DEVELOPMENT) && activeProfiles
-        .contains(JHipsterConstants.SPRING_PROFILE_PRODUCTION)) {
+    if (activeProfiles.contains("dev") && activeProfiles.contains("prod")) {
       log.error("You have misconfigured your application! It should not run " +
           "with both the 'dev' and 'prod' profiles at the same time.");
     }
-    if (activeProfiles.contains(JHipsterConstants.SPRING_PROFILE_DEVELOPMENT) && activeProfiles
-        .contains(JHipsterConstants.SPRING_PROFILE_CLOUD)) {
+    if (activeProfiles.contains("dev") && activeProfiles.contains("cloud")) {
       log.error("You have misconfigured your application! It should not" +
           "run with both the 'dev' and 'cloud' profiles at the same time.");
     }
