@@ -1,6 +1,5 @@
 package com.transformuk.hee.tis.profile.aop.logging;
 
-import io.github.jhipster.config.JHipsterConstants;
 import java.util.Arrays;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -11,6 +10,7 @@ import org.aspectj.lang.annotation.Pointcut;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.env.Environment;
+import org.springframework.core.env.Profiles;
 
 /**
  * Aspect for logging execution of service and repository Spring components.
@@ -41,8 +41,8 @@ public class LoggingAspect {
    */
   @AfterThrowing(pointcut = "loggingPointcut()", throwing = "e")
   public void logAfterThrowing(JoinPoint joinPoint, Throwable e) {
-    if (env.acceptsProfiles(JHipsterConstants.SPRING_PROFILE_DEVELOPMENT)) {
-      log.error("Exception in {}.{}() with cause = \'{}\' and exception = \'{}\'",
+    if (env.acceptsProfiles(Profiles.of("dev"))) {
+      log.error("Exception in {}.{}() with cause = '{}' and exception = '{}'",
           joinPoint.getSignature().getDeclaringTypeName(),
           joinPoint.getSignature().getName(), e.getCause() != null ? e.getCause() : "NULL",
           e.getMessage(), e);
