@@ -1,29 +1,12 @@
 package com.transformuk.hee.tis.profile.config;
 
-import io.github.jhipster.config.locale.AngularCookieLocaleResolver;
-import org.springframework.context.EnvironmentAware;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
 
 @Configuration
-public class LocaleConfiguration extends WebMvcConfigurerAdapter implements EnvironmentAware {
-
-  @Override
-  public void setEnvironment(Environment environment) {
-    // unused
-  }
-
-  @Bean(name = "localeResolver")
-  public LocaleResolver localeResolver() {
-    AngularCookieLocaleResolver cookieLocaleResolver = new AngularCookieLocaleResolver();
-    cookieLocaleResolver.setCookieName("NG_TRANSLATE_LANG_KEY");
-    return cookieLocaleResolver;
-  }
+public class LocaleConfiguration implements WebMvcConfigurer {
 
   @Override
   public void addInterceptors(InterceptorRegistry registry) {
