@@ -1,106 +1,56 @@
 package com.transformuk.hee.tis.profile.config;
 
-import com.codahale.metrics.MetricRegistry;
-import com.codahale.metrics.servlet.InstrumentedFilter;
-import com.codahale.metrics.servlets.MetricsServlet;
-import io.github.jhipster.config.JHipsterConstants;
-import io.github.jhipster.config.JHipsterProperties;
-import io.github.jhipster.web.filter.CachingHttpHeadersFilter;
 import io.undertow.UndertowOptions;
 import java.io.File;
 import java.nio.file.Paths;
-import java.util.EnumSet;
-import javax.servlet.DispatcherType;
-import javax.servlet.FilterRegistration;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
-import org.springframework.boot.context.embedded.MimeMappings;
-import org.springframework.boot.context.embedded.undertow.UndertowEmbeddedServletContainerFactory;
-import org.springframework.boot.web.servlet.ServletContextInitializer;
-import org.springframework.context.annotation.Bean;
+import org.springframework.boot.web.embedded.undertow.UndertowServletWebServerFactory;
+import org.springframework.boot.web.server.MimeMappings;
+import org.springframework.boot.web.server.WebServerFactoryCustomizer;
+import org.springframework.boot.web.servlet.server.ConfigurableServletWebServerFactory;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
 
 /**
  * Configuration of web application with Servlet 3.0 APIs.
  */
 @Configuration
-public class WebConfigurer implements ServletContextInitializer,
-    EmbeddedServletContainerCustomizer {
+public class WebConfigurer implements WebServerFactoryCustomizer<ConfigurableServletWebServerFactory> {
 
   private final Logger log = LoggerFactory.getLogger(WebConfigurer.class);
-
-  private final Environment env;
-
-  private final JHipsterProperties jHipsterProperties;
-
-  private MetricRegistry metricRegistry;
-
-  public WebConfigurer(Environment env, JHipsterProperties jHipsterProperties) {
-
-    this.env = env;
-    this.jHipsterProperties = jHipsterProperties;
-  }
-
-  @Override
-  public void onStartup(ServletContext servletContext) throws ServletException {
-    if (env.getActiveProfiles().length != 0) {
-      log.info("Web application configuration, using profiles: {}",
-          (Object[]) env.getActiveProfiles());
-    }
-    EnumSet<DispatcherType> disps = EnumSet
-        .of(DispatcherType.REQUEST, DispatcherType.FORWARD, DispatcherType.ASYNC);
-    initMetrics(servletContext, disps);
-    if (env.acceptsProfiles(JHipsterConstants.SPRING_PROFILE_PRODUCTION)) {
-      initCachingHttpHeadersFilter(servletContext, disps);
-    }
-    log.info("Web application fully configured");
-  }
 
   /**
    * Customize the Servlet engine: Mime types, the document root, the cache.
    */
   @Override
-  public void customize(ConfigurableEmbeddedServletContainer container) {
+  public void customize(ConfigurableServletWebServerFactory webServerFactory) {
     MimeMappings mappings = new MimeMappings(MimeMappings.DEFAULT);
     // IE issue, see https://github.com/jhipster/generator-jhipster/pull/711
     mappings.add("html", "text/html;charset=utf-8");
     // CloudFoundry issue, see https://github.com/cloudfoundry/gorouter/issues/64
     mappings.add("json", "text/html;charset=utf-8");
-    container.setMimeMappings(mappings);
+    webServerFactory.setMimeMappings(mappings);
     // When running in an IDE or with ./mvnw spring-boot:run, set location of the static web assets.
-    setLocationForStaticAssets(container);
+    setLocationForStaticAssets(webServerFactory);
 
     /*
      * Enable HTTP/2 for Undertow - https://twitter.com/ankinson/status/829256167700492288
      * HTTP/2 requires HTTPS, so HTTP requests will fallback to HTTP/1.1.
-     * See the JHipsterProperties class and your application-*.yml configuration files
-     * for more information.
      */
-    if (jHipsterProperties.getHttp().getVersion().equals(JHipsterProperties.Http.Version.V_2_0) &&
-        container instanceof UndertowEmbeddedServletContainerFactory) {
+    if (webServerFactory instanceof UndertowServletWebServerFactory) {
 
-      ((UndertowEmbeddedServletContainerFactory) container)
-          .addBuilderCustomizers(builder ->
-              builder.setServerOption(UndertowOptions.ENABLE_HTTP2, true));
+      ((UndertowServletWebServerFactory) webServerFactory)
+              .addBuilderCustomizers(builder ->
+                      builder.setServerOption(UndertowOptions.ENABLE_HTTP2, true));
     }
   }
 
-  private void setLocationForStaticAssets(ConfigurableEmbeddedServletContainer container) {
+  private void setLocationForStaticAssets(ConfigurableServletWebServerFactory webServerFactory) {
     File root;
     String prefixPath = resolvePathPrefix();
     root = new File(prefixPath + "ui-build/profile/");
     if (root.exists() && root.isDirectory()) {
-      container.setDocumentRoot(root);
+      webServerFactory.setDocumentRoot(root);
     }
   }
 
@@ -116,63 +66,5 @@ public class WebConfigurer implements ServletContextInitializer,
       return "";
     }
     return extractedPath.substring(0, extractionEndIndex);
-  }
-
-  /**
-   * Initializes the caching HTTP Headers Filter.
-   */
-  private void initCachingHttpHeadersFilter(ServletContext servletContext,
-      EnumSet<DispatcherType> disps) {
-    log.debug("Registering Caching HTTP Headers Filter");
-    FilterRegistration.Dynamic cachingHttpHeadersFilter =
-        servletContext.addFilter("cachingHttpHeadersFilter",
-            new CachingHttpHeadersFilter(jHipsterProperties));
-
-    cachingHttpHeadersFilter.addMappingForUrlPatterns(disps, true, "/content/*");
-    cachingHttpHeadersFilter.addMappingForUrlPatterns(disps, true, "/app/*");
-    cachingHttpHeadersFilter.setAsyncSupported(true);
-  }
-
-  /**
-   * Initializes Metrics.
-   */
-  private void initMetrics(ServletContext servletContext, EnumSet<DispatcherType> disps) {
-    log.debug("Initializing Metrics registries");
-    servletContext.setAttribute(InstrumentedFilter.REGISTRY_ATTRIBUTE,
-        metricRegistry);
-    servletContext.setAttribute(MetricsServlet.METRICS_REGISTRY,
-        metricRegistry);
-
-    log.debug("Registering Metrics Filter");
-    FilterRegistration.Dynamic metricsFilter = servletContext.addFilter("webappMetricsFilter",
-        new InstrumentedFilter());
-
-    metricsFilter.addMappingForUrlPatterns(disps, true, "/*");
-    metricsFilter.setAsyncSupported(true);
-
-    log.debug("Registering Metrics Servlet");
-    ServletRegistration.Dynamic metricsAdminServlet =
-        servletContext.addServlet("metricsServlet", new MetricsServlet());
-
-    metricsAdminServlet.addMapping("/management/metrics/*");
-    metricsAdminServlet.setAsyncSupported(true);
-    metricsAdminServlet.setLoadOnStartup(2);
-  }
-
-  @Bean
-  public CorsFilter corsFilter() {
-    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-    CorsConfiguration config = jHipsterProperties.getCors();
-    if (config.getAllowedOrigins() != null && !config.getAllowedOrigins().isEmpty()) {
-      log.debug("Registering CORS filter");
-      source.registerCorsConfiguration("/api/**", config);
-      source.registerCorsConfiguration("/v2/api-docs", config);
-    }
-    return new CorsFilter(source);
-  }
-
-  @Autowired(required = false)
-  public void setMetricRegistry(MetricRegistry metricRegistry) {
-    this.metricRegistry = metricRegistry;
   }
 }
