@@ -3,19 +3,20 @@ package com.transformuk.hee.tis.profile.service;
 import static com.google.common.collect.Lists.newArrayList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Matchers.anyListOf;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import com.transformuk.hee.tis.profile.domain.TraineeProfile;
 import com.transformuk.hee.tis.profile.dto.RegistrationRequest;
 import com.transformuk.hee.tis.profile.repository.TraineeProfileRepository;
+import java.util.Collection;
 import java.util.List;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -44,8 +45,6 @@ public class TraineeProfileServiceTest {
     existingTraineeProfile.setDesignatedBodyCode(DBC);
 
     // given
-    given(traineeProfileRepository.findByDesignatedBodyCode(DBC))
-        .willReturn(newArrayList(existingTraineeProfile));
     given(traineeProfileRepository.findByGmcNumberIn(gmcNumbers))
         .willReturn(newArrayList(existingTraineeProfile));
 
@@ -58,30 +57,39 @@ public class TraineeProfileServiceTest {
   }
 
   @Test
+  @SuppressWarnings("unchecked")
   public void shouldCreateIfNotExists() {
-    RegistrationRequest request = new RegistrationRequest();
-    request.setGmcNumber(EXISTING_GMC_NUMBER);
+    RegistrationRequest moved = new RegistrationRequest();
+    moved.setGmcNumber(EXISTING_GMC_NUMBER);
+    RegistrationRequest brandNew = new RegistrationRequest();
+    brandNew.setGmcNumber(NEW_GMC_NUMBER);
 
-    List<String> gmcNumbers = newArrayList(EXISTING_GMC_NUMBER, NEW_GMC_NUMBER);
     TraineeProfile existingTraineeProfile = new TraineeProfile(1L, EXISTING_GMC_NUMBER);
 
     TraineeProfile newTraineeProfile = new TraineeProfile(null, NEW_GMC_NUMBER);
     // given
-    given(traineeProfileRepository.findByGmcNumberIn(gmcNumbers))
+    ArgumentCaptor<Collection<String>> stringsCaptor = ArgumentCaptor.forClass(Collection.class);
+    given(traineeProfileRepository.findByGmcNumberIn(stringsCaptor.capture()))
         .willReturn(newArrayList(existingTraineeProfile));
-    given(traineeProfileRepository.save(anyListOf(TraineeProfile.class))).willReturn(newArrayList
-        (existingTraineeProfile, newTraineeProfile));
+    ArgumentCaptor<Iterable<TraineeProfile>> actualProfileCaptor =
+        ArgumentCaptor.forClass(Iterable.class);
+    given(traineeProfileRepository.saveAll(actualProfileCaptor.capture()))
+        .willReturn(newArrayList(newTraineeProfile))
+        .willReturn(newArrayList(existingTraineeProfile));
 
     // when
-    List<TraineeProfile> traineeProfiles = service.findOrCreate(DBC, newArrayList(request));
+    List<TraineeProfile> actualProfiles = service.findOrCreate(DBC, newArrayList(moved, brandNew));
 
     // then
-    assertThat(traineeProfiles).isEqualTo(newArrayList(existingTraineeProfile, newTraineeProfile));
+    assertThat(actualProfiles).containsExactlyInAnyOrder(existingTraineeProfile, newTraineeProfile);
+    assertThat(stringsCaptor.getValue())
+        .containsExactlyInAnyOrder(EXISTING_GMC_NUMBER, NEW_GMC_NUMBER);
+    assertThat(actualProfileCaptor.getAllValues().get(0)).containsOnlyOnce(newTraineeProfile);
   }
 
   @Test
   public void shouldReturnMappings() {
-    Pageable pageable = new PageRequest(0, 100);
+    Pageable pageable = PageRequest.of(0, 100);
     TraineeProfile existingTraineeProfile = new TraineeProfile(1L, EXISTING_GMC_NUMBER);
     Page<TraineeProfile> page = new PageImpl<>(newArrayList(existingTraineeProfile));
     // given
@@ -93,5 +101,4 @@ public class TraineeProfileServiceTest {
     // then
     assertThat(actualPage).isSameAs(page);
   }
-
 }
